wxOSX: activate app on LSUIElement before bundle check
This patch is useful to wx deployments where we have LSUIElement
(and thus want the latest fix with the app activation) but where
the program doesn't appear to the system to be running in the bundle,
thus failing a check like `bundleUrl.EndsWith(".app")`.

The bundle check is not necessary because if LSUIElement was set we're
definitely in a bundle.

The bundle check is specifically problem for wxErlang where we have a
bundle like this:

    MyApp.app/
      Contents/
        Info.plist
        MacOS/
          MyApp # an executable/script that launches ../Resources/erlang/bin/erlexec
        Resources/
          erlang/
            erts/bin/
              erlexec  # execv beam.smp
              beam.smp # this is the "final" executable
              ...
            lib/
            ...

But then even though the application _is_ in a bundle and things like
LSUIElement are respected, the returned bundle URL is _wrong_, it
basically gives the directory where the "final" executable is. Thus the
`bundleUrl.EndsWith(".app")` check fails,

    [NSApp setActivationPolicy: NSApplicationActivationPolicyRegular];

is called, and that is definitely not what we want with LSUIElement,
as it overrides it.